fix broken authnStore implementation
