+++ frontend/src/store/authn/helper.js
@@ -1,0 +1,228 @@
+//
+// SPDX-FileCopyrightText: 2023 SAP SE or an SAP affiliate company and Gardener contributors
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+
+import { unref } from 'vue'
+import { useBrowserLocation } from '@vueuse/core'
+import { useCookies } from '@vueuse/integrations/useCookies'
+import decode from 'jwt-decode'
+
+import { useLogger } from '@/composables/useLogger'
+
+import {
+  createError,
+  createNoUserError,
+  createClockSkewError,
+  isUnauthorizedError,
+  isNoUserError,
+  isClockSkewError,
+} from '@/utils/errors'
+
+export const COOKIE_HEADER_PAYLOAD = 'gHdrPyl'
+const CLOCK_TOLERANCE = 15
+
+function now () {
+  return Math.floor(Date.now() / 1000)
+}
+
+function delay (duration = 0) {
+  return new Promise(resolve => setTimeout(resolve, duration))
+}
+
+function secondsUntil (val) {
+  if (val) {
+    const t = Number(val)
+    if (!Number.isNaN(t)) {
+      return t - now()
+    }
+  }
+}
+
+function isRefreshRequired (user = {}) {
+  if (!user.rti || !user.refresh_at) {
+    return false
+  }
+  const t = secondsUntil(user.refresh_at)
+  return typeof t === 'number' && t < CLOCK_TOLERANCE
+}
+
+function isExpired (user = {}) {
+  if (!user.exp) {
+    return true
+  }
+  const t = secondsUntil(user.exp)
+  return typeof t === 'number' && t < CLOCK_TOLERANCE
+}
+
+export function useUserManager (options) {
+  const {
+    cookies = useCookies([COOKIE_HEADER_PAYLOAD]),
+    location = useBrowserLocation(),
+    logger = useLogger(),
+  } = options ?? {}
+
+  let refreshTokenPromise
+
+  const origin = unref(location).origin
+
+  function decodeCookie () {
+    try {
+      const value = cookies.get(COOKIE_HEADER_PAYLOAD)
+      if (value) {
+        return decode(value)
+      }
+    } catch (err) {
+      logger.error(err.message)
+    }
+    return null
+  }
+
+  function deleteCookie () {
+    cookies.remove(COOKIE_HEADER_PAYLOAD)
+  }
+
+  async function createTokenRefreshRequest () {
+    const timestamp = Math.floor(Date.now() / 1000)
+    const response = await fetch('/auth/token', {
+      method: 'POST',
+      cache: 'no-store',
+      headers: {
+        accept: 'application/json',
+        'x-requested-with': 'XMLHttpRequest',
+        'content-type': 'application/json',
+      },
+      body: JSON.stringify({ timestamp }),
+    })
+
+    const statusCode = response.status
+    if (statusCode >= 200 && statusCode < 300) {
+      const user = decodeCookie()
+      if (!user) {
+        throw createNoUserError()
+      }
+      if (isRefreshRequired(user)) {
+        throw createClockSkewError()
+      }
+      return user
+    }
+
+    if (statusCode >= 400) {
+      let message = `Token refresh failed with status code ${statusCode}`
+      try {
+        const data = await response.json()
+        Object.defineProperty(response, 'data', { value: data })
+        if (data.message) {
+          message = data.message
+          if (statusCode === 401) {
+            /*
+             * The OPError message has the following format `${error} (${error_description})`
+             * (see https://github.com/panva/node-openid-client/blob/1e3892e6222fdd1956735f97584ebc722fcebdd3/lib/errors.js#L5)
+             * The error and the error_description are values returned in the error response
+             * from the the OpenID Connect Provider (OP). We use the original OP error_description
+             * as error message if possible.
+             */
+            const matches = /^(.+) \((.+)\)$/.exec(message)
+            if (matches && matches.length > 2) {
+              message = matches[2]
+            }
+          }
+        }
+      } catch (err) { /* ignore error */ }
+      throw createError(statusCode, message, { response })
+    }
+
+    throw createError(statusCode, 'Token refresh failed', { response })
+  }
+
+  function redirect (url) {
+    unref(location).href = url
+  }
+
+  function signout (err) {
+    deleteCookie()
+    const url = new URL('/auth/logout', origin)
+    if (err) {
+      url.searchParams.set('error[name]', err.name)
+      url.searchParams.set('error[message]', err.message)
+    }
+    redirect(url)
+  }
+
+  function signin () {
+    const url = new URL('/login', origin)
+    redirect(url)
+  }
+
+  function signinWithOidc (redirectPath = '/') {
+    const url = new URL('/auth', origin)
+    const redirectUrl = new URL(redirectPath, origin)
+    url.searchParams.set('redirectUrl', redirectUrl)
+    redirect(url)
+  }
+
+  async function refreshToken () {
+    try {
+      let user = decodeCookie()
+      if (!user) {
+        throw createNoUserError()
+      }
+      if (!isRefreshRequired(user)) {
+        return
+      }
+      logger.debug('Acquiring token refresh lock (%s)', user.rti)
+      await navigator.locks.request('token-refresh-request', async () => {
+        user = decodeCookie()
+        if (!user) {
+          throw createNoUserError()
+        }
+        if (!isRefreshRequired(user)) {
+          return
+        }
+        logger.debug('Refreshing token (%s)', user.rti)
+        const oldUser = user
+        user = await createTokenRefreshRequest()
+        logger.debug('Token has been refreshed (%s <- %s)', user.rti, oldUser.rti)
+      })
+    } catch (err) {
+      logger.error('Token refresh failed: %s - %s', err.name, err.message)
+      let frameRequestCallback
+      if (isNoUserError(err)) {
+        frameRequestCallback = () => signin()
+      } else if (isUnauthorizedError(err) || isClockSkewError(err)) {
+        frameRequestCallback = () => signout(err)
+      }
+      if (typeof frameRequestCallback === 'function') {
+        window.requestAnimationFrame(frameRequestCallback)
+        // delay the error by 500 ms to avoid rendering or navigation before redirection has started
+        await delay(500)
+      }
+      throw err
+    }
+  }
+
+  function ensureValidToken () {
+    if (!refreshTokenPromise) {
+      refreshTokenPromise = refreshToken().finally(() => {
+        refreshTokenPromise = undefined
+      })
+    }
+    return refreshTokenPromise
+  }
+
+  return {
+    decodeCookie,
+    isRefreshRequired () {
+      return isRefreshRequired(decodeCookie())
+    },
+    isExpired () {
+      return isExpired(decodeCookie())
+    },
+    signin,
+    signout,
+    signinWithOidc,
+    refreshToken,
+    ensureValidToken,
+  }
+}
